made program to determine the amount of change due - needs work deleting the decimals

+++ PycharmProjects/untitled/Change.py
@@ -1,7 +1,7 @@
 cents = raw_input("Enter a whole number from 1 to 99.\nThe machine will"
                   " determine a combination of coins.\n")
 print cents, " cents in coins:"
-cents = float(cents)
+cents = int(cents)
 numQ = cents/25
 cents = cents % 25
 numD = cents/10
